requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
numpy==1.26.4
scikit-learn==1.5.2
python-docx==1.1.2
docxtpl==0.16.7
langid==1.1.6
pymupdf==1.24.9
openpyxl==3.1.5
python-pptx==0.6.23
openai==1.47.0
email-validator>=2
Pillow>=10
python-docx
docxtpl
lxml
azure-cosmos
azure-storage-blob
python-jose[cryptography]==3.3.0
